make executor config pflags-generatable and regenerate
The pflags generator does not support uint32, so config_flags.go could not be
regenerated and the prior maxSystemFailures flag entry was hand-added. Change
Config.MaxSystemFailures from uint32 to int32 (cast back to uint32 at the
reconciler boundary, where it is compared against Status.SystemFailures uint32)
and run `go generate ./executor/pkg/config/...`. config_flags.go and
config_flags_test.go are now genuine generator output, including the new
defaultK8sServiceAccount flag.

+++ executor/setup.go
@@ -170,7 +170,7 @@ func Setup(ctx context.Context, sc *app.SetupContext) error {
 	reconciler.CatalogClient = asyncCatalogClient
 	reconciler.Catalog = cacheClient
 	reconciler.Recorder = mgr.GetEventRecorderFor("taskaction-controller")
-	reconciler.MaxSystemFailures = cfg.MaxSystemFailures
+	reconciler.MaxSystemFailures = uint32(cfg.MaxSystemFailures)
 	if err := reconciler.SetupWithManager(mgr); err != nil {
 		return fmt.Errorf("executor: failed to setup controller: %w", err)
 	}
